# resources/SEQENS_Swagger-v1.0.6.yml
openapi: 3.0.1
info:
  title: SEQENS - Datahub
  description: API collection converted from Postman
  version: 1.0.6
servers:
  - url: "https://env11.ivalua.app/buyer/actionlogement/devmaint2/7wbj8"
    description: Development Environment
  - url: "https://env03.ivalua.com/buyer/actionlogement/rctmaint2/yfwag"
    description: Recette Environment
  - url: "https://env21.ivalua.app/buyer/actionlogement/sandboxmaint/rvr4t/"
    description: Sandbox Environment
...
paths:
...
  /api.aspx/v1.0/ord/orders:
    get:
      tags:
        - API Commandes
      summary: Récupération des commandes
      description:
        Récupère la liste des commandes en mode `full` (tous les Commandes) ou `diff` (commandes modifiés selon critères).
        Le format de réponse peut être `xml` ou `json`.
      parameters:
        - name: format
          in: query
          required: true
          description: "Format de réponse souhaité. Valeurs possibles : `xml` ou `json`."
          schema:
            $ref: "#/components/schemas/FormatEnum"

        - name: mode
          in: query
          required: true
          description: "Mode de récupération des commandes. Valeurs possibles: `full` ou `diff`."
          schema:
            $ref: "#/components/schemas/ModeEnum"
            
        - name: ord_id
          in: query
          required: false
          description: Identifiant de la commande (mode diff uniquement, plusieurs commandes possibles séparées par des virgules)
          schema:
            type: string
            
        - name: sup_id
          in: query
          required: false
          description: Référence du fournisseur (mode diff uniquement, plusieurs fournisseurs possibles séparées par des virgules)
          schema:
            type: string
            
        - name: tiers_code
          in: query
          required: false
          description: Identitfiant national du fournisseur (mode diff uniquement)
          schema:
            type: string
            
        - name: date_from
          in: query
          required: false
          description: Date de début pour filtrer les Commandes (mode diff uniquement)
          schema:
            type: string
            format: date
            
        - name: date_to
          in: query
          required: false
          description: Date de fin pour filtrer les Commandes (mode diff uniquement)
          schema:
            type: string
            format: date
            
      responses:
        "200":
          description: Liste des Commandes récupérée avec succès
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
        "400":
          description: Erreur de validation des paramètres d'entrée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersErrorResponse"
  /api.aspx/v1.0/ord/orders/{object_id}:
    get:
      tags:
        - API Commandes
      summary: Récupération des Commandes
      description:
        Récupère la liste des Commandes en mode `full` (tous les Commandes) ou `diff` (Commandes modifiés selon critères).
        Le format de réponse peut être `xml` ou `json`.
      parameters:
        - name: object_id
          in: path
          required: true
          schema:
            type: string
          description: "Un ID d'Objet peut être déclaré avec Path comme origine."
          
        - name: format
          in: query
          required: true
          description: "Format de réponse souhaité. Valeurs possibles : `xml` ou `json`."
          schema:
            $ref: "#/components/schemas/FormatEnum"
            
        - name: mode
          in: query
          required: true
          description: "Mode de récupération des commandes. Valeurs possibles: `full` ou `diff`."
          schema:
            $ref: "#/components/schemas/ModeEnum"
            
        - name: ord_id
          in: query
          required: false
          description: Identifiant de la commande (mode diff uniquement, plusieurs commandes possibles séparées par des virgules)
          schema:
            type: integer
            
        - name: sup_id
          in: query
          required: false
          description: Référence du fournisseur (mode diff uniquement, plusieurs fournisseurs possibles séparées par des virgules)
          schema:
            type: string
            
        - name: ctr_code
          in: query
          required: false
          description: Code du contrat (mode diff uniquement)
          schema:
            type: string
            
        - name: tiers_code
          in: query
          required: false
          description: Code fournisseur (mode diff uniquement)
          schema:
            type: string
            
        - name: date_from
          in: query
          required: false
          description: Date de début pour filtrer les Commandes (mode diff uniquement)
          schema:
            type: string
            format: date
            
        - name: date_to
          in: query
          required: false
          description: Date de fin pour filtrer les Commandes (mode diff uniquement)
          schema:
            type: string
            format: date
            
      responses:
        "200":
          description: Liste des Commandes récupérée avec succès
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
        "400":
          description: Erreur de validation des paramètres d'entrée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersErrorResponse"
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://env21.ivalua.app/buyer/actionlogement/sandboxmaint/rvr4t/oauth2/token
          scopes:
            auth_api_api_endpoint_exec: "Accès aux API de l'exécution"
    apikeyAuth:
      type: apiKey
      in: header
      name: apikey
    apikeyAuthQuery:
      type: apiKey
      in: query
      name: apikey
  schemas:
    FormatEnum:
      type: string
      enum: [json, xml]
      description: "Format des données. Valeurs possibles : `json` ou `xml`."
    ModeEnum:
      type: string
      enum: [full, diff]
      description: "Mode de récupération. Valeurs possibles : `full` ou `diff`."
      
    # API Fournisseur
...
    OrdersResponse:
      type: object
      properties:
        header:
          type: object
          properties:
            apiName:
              type: string
              example: "Orders"
            format:
              type: string
              example: "json"
            totalRow:
              type: integer
              example: 1
        orders:
          type: array
          items:
            type: object
            properties:
              dataOrder:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  objectId:
                    type: integer
                    example: 1
                  ordIdOrigin:
                    type: integer
                    example: 1
                  orderCode:
                    type: string
                    example: "PO000001"
                  orderLabel:
                    type: string
                    example: "D.A. du 21/03/2025"
                  ordTypeCode:
                    type: string
                    example: "default"
                  ordExtCode:
                    type: string
                    example: ""
                  ordRef:
                    type: string
                    example: ""
                  basketId:
                    type: integer
                    example: 5
                  orderSupId:
                    type: integer
                    example: 3
                  orderSupName:
                    type: string
                    example: "TEST_COMP"
                  supNatId:
                    type: string
                    example: ""
                  supNatIdType:
                    type: string
                    example: ""
                  created:
                    type: string
                    format: date
                    example: "2025-03-21"
                  modified:
                    type: string
                    format: date
                    example: "2025-03-21"
                  loginCreated:
                    type: string
                    example: "JohnD"
                  loginModified:
                    type: string
                    example: "JohnD"
                  statusCode:
                    type: string
                    example: "end"
                  statusLabel:
                    type: string
                    example: "Terminée"
                  ordOrderDate:
                    type: string
                    format: date
                    example: "2025-03-21"
                  oitemsTotalAmount:
                    type: number
                    format: float
                    example: 120000.00
                  unitCodeCurrency:
                    type: string
                    example: "EUR"
                  ordComment:
                    type: string
                    example: ""
                  incoCode:
                    type: string
                    example: ""
                  ordIncoPlace:
                    type: string
                    example: ""
                  paytermCode:
                    type: string
                    example: "10"
                  paytermLabel:
                    type: string
                    example: "30 jours date de facture"
                  paymentTypeCode:
                    type: string
                    example: ""
                  paymentTypeLabel:
                    type: string
                    example: ""
                  ordFreeBudget:
                    type: string
                    example: "No"
                  ordAmendmentNum:
                    type: string
                    example: "0"
                  ordTrackTimeSheet:
                    type: string
                    example: "0"
                  legalCompCode:
                    type: string
                    example: "V90"
                  legalCompLegalForm:
                    type: string
                    example: ""
                  legalCompLabel:
                    type: string
                    example: "Seqens"
                  orgaLevel:
                    type: string
                    example: "site"
                  orgaNode:
                    type: string
                    example: "SEM_DD20"
                  orgaLabel:
                    type: string
                    example: "DD Hauts-de-Seine"
              orderContacts:
                type: array
                items:
                  type: object
                  properties:
                    contactRequesterFirstname:
                      type: string
                      example: "John"
                    contactRequesterLastname:
                      type: string
                      example: "DOE"
                    contactRequesterEmail:
                      type: string
                      format: email
                      example: "john.doe@test.eu"
                    contactBillingFirstname:
                      type: string
                      example: ""
                    contactBillingLastname:
                      type: string
                      example: ""
                    contactBillingEmail:
                      type: string
                      format: email
                      example: ""
                    contactDeliveryFirstname:
                      type: string
                      example: ""
                    contactDeliveryLastname:
                      type: string
                      example: ""
                    contactDeliveryEmail:
                      type: string
                      format: email
                      example: ""
                    contactSupplierFirstname:
                      type: string
                      example: "Malcom"
                    contactSupplierLastname:
                      type: string
                      example: "X"
                    contactSupplierEmail:
                      type: string
                      format: email
                      example: "Malc.x@test.com"
              orderItems:
                type: array
                items:
                  type: object
                  properties:
                    oitemId:
                      type: integer
                      example: 8
                    oitemLabel:
                      type: string
                      example: "Forfait annuel Maintenance Ascenseurs HP2"
                    oitemFamLabel:
                      type: string
                      example: "Ascenseurs"
                    oitemFamNode:
                      type: string
                      example: "325"
                    oitemFamLevel:
                      type: string
                      example: "ssfam"
                    oitemQuantity:
                      type: number
                      format: float
                      example: 1
                    oitemTotalAmount:
                      type: number
                      format: float
                      example: 120000.00
              addresses:
                type: array
                items:
                  type: object
                  properties:
                    type:
                      type: string
                      enum: [billing, delivery]
                      example: billing
                    adrNum:
                      type: string
                      example: ""
                    adrNomComplt:
                      type: string
                      example: ""
                    adrVoie:
                      type: string
                      example: "Bd Garibaldi 14-16"
                    adrVoieComplt:
                      type: string
                      example: ""
                    zipCode:
                      type: string
                      example: "92130"
                    zipLabel:
                      type: string
                      example: "Issy-les-Moulineaux"
                    countryCode:
                      type: string
                      example: "FR"
                    countryLabel:
                      type: string
                      example: "FRANCE"
    OrdersErrorResponse:
      type: object
      properties:
        erreurs:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                example: "ERR-QUE-001"
              message:
                type: string
                example: "Le mode doit être renseigné."
      example:
        erreurs:
          - code: "ERR-QUE-101"
            message: "Le mode doit être renseigné."
          - code: "ERR-QUE-102"
            message: "Le mode doit être soit 'full', soit 'diff'."
          - code: "ERR-QUE-103"
            message: "Le format spécifié doit être 'json' , 'xml' ou 'scim'."
          - code: "ERR-QUE-104"
            message: "Erreur : Au moins un paramètre doit être fourni."
          - code: "ERR-QUE-105"
            message: "Une date (date_from ou date_to) est manquante."
          - code: "ERR-QUE-106"
            message: "Le paramètre date_from doit être inférieur à date_to."
          - code: "ERR-QUE-107"
            message: "L’intervalle entre date_from et date_to ne peut pas dépasser 2 mois."
            
    # API Factures
...